requirements.txt
bson==0.5.10
click==7.1.2
Flask==1.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pymongo==3.11.3
python-dateutil==2.8.1
python-dotenv==0.17.0
six==1.15.0
Werkzeug==1.0.1
